+++ server/message/parts/CustomMessage.jsx
@@ -1,5 +1,6 @@
 /** @jsx h */
-import { h } from 'preact';
+/** @jsxFrag Fragment */
+import { h, Fragment } from 'preact';
 import render from 'preact-render-to-string';
 import { getLogos } from '../../locale';
 import { getDataByTag } from '../../../src/utils/server';
@@ -32,7 +33,8 @@ const CustomMessage = ({ children, data, meta, template }) => {
         const [type, ...parts] = templateVariable.split('.');
 
         if (type === 'logo') {
-            const src = getLogos(meta.offerCountry)[parts[0].toUpperCase()][parts[1].toUpperCase()]?.src;
+            const src = getLogos(meta.offerCountry, meta.offerType)[parts[0].toUpperCase()][parts[1].toUpperCase()]
+                ?.src;
             return `<img alt="PayPal Credit logo" src="${src}" />`;
         }
 
@@ -45,11 +47,11 @@ const CustomMessage = ({ children, data, meta, template }) => {
         );
     });
     return (
-        <div role="button" className="message" tabIndex="0">
+        <>
             {children}
-            {/* eslint-disable-next-line react/no-danger */}
-            <div dangerouslySetInnerHTML={{ __html: populatedMarkup }} />
-        </div>
+            {/* eslint-disable-next-line react/no-danger, jsx-a11y/control-has-associated-label */}
+            <div role="button" className="message" tabIndex="0" dangerouslySetInnerHTML={{ __html: populatedMarkup }} />
+        </>
     );
 };