+++ src/utils/stats.js
@@ -1,7 +1,8 @@
+import { SDK_SETTINGS } from '@paypal/sdk-constants';
 import { checkAdblock } from './adblock';
-import { isHidden, isInViewport } from './elements';
+import { isHidden, isInViewport, getTopWindow } from './elements';
 import { logger } from './logger';
-import { getLibraryVersion } from './sdk';
+import { getLibraryVersion, getScriptAttributes } from './sdk';
 
 const scrollHandlers = new Map();
 const handleScroll = event => scrollHandlers.forEach(handler => handler(event));
@@ -22,6 +23,9 @@ const onScroll = (elem, handler) => {
 export function runStats({ container, activeTags, index }) {
     // Get outer most container's page location coordinates
     const containerRect = container.getBoundingClientRect();
+    const topWindow = getTopWindow();
+
+    const sdkMetaAttributes = getScriptAttributes();
 
     // Create initial payload
     const payload = {
@@ -30,12 +34,13 @@ export function runStats({ container, activeTags, index }) {
         event_type: 'stats',
         integration_type: __MESSAGES__.__TARGET__,
         messaging_version: getLibraryVersion(),
+        bn_code: sdkMetaAttributes[SDK_SETTINGS.PARTNER_ATTRIBUTION_ID],
         // Beaver logger filters payload props based on Boolean conversion value
         // so everything must be converted to a string to prevent unintended filtering
         pos_x: Math.round(containerRect.left).toString(),
         pos_y: Math.round(containerRect.top).toString(),
-        browser_width: window.innerWidth.toString(),
-        browser_height: window.innerHeight.toString(),
+        browser_width: (topWindow?.innerWidth).toString(),
+        browser_height: (topWindow?.innerHeight).toString(),
         visible: isInViewport(container).toString(),
         // Visible message sections
         active_tags: activeTags