+++ tests/functional/spec/GB/modal/gb_modalTestDefs.js
@@ -1,4 +1,5 @@
 import selectors from '../../utils/selectors';
+import logTestName from '../../utils/logTestName';
 import modalSnapshot from '../../utils/modalSnapshot';
 
 /**
@@ -7,15 +8,17 @@ import modalSnapshot from '../../utils/modalSnapshot';
 
 export const gbModalContent = ({ account, viewport, groupString }) => async () => {
     const testNameParts = 'gb modal content';
-    const elementModal = await page.$('iframe[title*="paypal_credit_modal"]');
+    logTestName({ account, viewport, groupString, testNameParts });
+
+    const elementModal = await page.$(selectors.modal.iframe);
 
     const modalFrame = await elementModal.contentFrame();
     await page.waitFor(2000);
-    await modalFrame.waitForSelector(selectors.modal.modalContent);
-    await modalFrame.waitForSelector(selectors.modal.modalMain);
-    await modalFrame.waitForSelector(selectors.modal.containerLeft);
-    await modalFrame.waitForSelector(selectors.modal.offer);
-    const h1 = await modalFrame.evaluate(() => document.querySelector('.content-body__offer').innerHTML);
+    await modalFrame.waitForSelector(selectors.modal.gbContent);
+    await modalFrame.waitForSelector(selectors.modal.gbMain);
+    await modalFrame.waitForSelector(selectors.modal.gbContainerLeft);
+    await modalFrame.waitForSelector(selectors.modal.gbOffer);
+    const h1 = await modalFrame.$eval(selectors.modal.gbOffer, element => element.innerHTML);
 
     if (account.includes('GBPLQ')) {
         expect(h1).toContain('3 interest-free payments of <br> £41,67 per month with Flex');